Convert SizeOptions.Size from int to long - Storage APIs use long instead of int

+++ java/azure-perfstress/src/main/java/com/azure/perfstress/CircularFlux.java
@@ -5,11 +5,11 @@ import java.nio.ByteBuffer;
 import reactor.core.publisher.Flux;
 
 public class CircularFlux {
-    public static Flux<ByteBuffer> create(ByteBuffer byteBuffer, int size) {
+    public static Flux<ByteBuffer> create(ByteBuffer byteBuffer, long size) {
         int remaining = byteBuffer.remaining();
         
-        int quotient = size / remaining;
-        int remainder = size % remaining;
+        int quotient = (int) size / remaining;
+        int remainder = (int) size % remaining;
 
         return Flux.range(0, quotient)
             .map(i -> byteBuffer.duplicate())
